+++ Changelog.md
@@ -1,6 +1,7 @@
 ## 0.1.1 (unreleased)
 
 - Slightly improve performance of grapheme cluster segmentation.
+- Fixes grapheme cluster segmentation of multiple consecutive regional flags.
 
 ## 0.1.0 (2022-11-03)